+++ taka-reverse/src/Navbar /index.css
@@ -1,47 +1,70 @@
-#landingPage{
-    background-image: url('/public/media/recycling.png');
+
+.Navbar-container{
+    display: flex;
+    justify-content: center;
+    justify-content: space-between;
+    align-items: center;
+    position: fixed;
+    height: 130px;
+    /* margin-left:15px; */
+    width: 100vw;
     background-color: rgb(41,67,76);
-    background-repeat: no-repeat;
-    background-position:  50% 50%;
-    background-size: 70%,70%;
-    height: 1000px;
-    width: 100%;
-    margin: 0;
-    top: 0;
-    padding-top: 60px;
+      
+  }
+
+.show-nav a{
+    text-decoration: none;
+    color: white;
+}
+.show-nav a:hover{
+    color: green;
+}
+
+.show-nav{
+    display: flex;
+    gap: 80px;
+    margin-right: 40px;
+    font-size: larger;
+    z-index: 1001;
 }
 
 p{
     color:white;
-    
 }
 
-button{
-    
-    background-color: white;
-    border-radius: 30px;
-    border: none;
-    padding: 15px 32px;
-    text-align: center;
-    display: inline-block;
-    font-size: 20px;  
-    margin-top: 380px;   
-}
-button a{
-    text-decoration: none;
-    color: black;
-}
-.Recycling-context{
-   text-align: center;
-   justify-content: center;
-   padding-top: 10%;
-}
 span{
-    font-size: 200px; 
-} 
+    font-size: 200px;  
+}
+.hamburger-menu{
+    flex-direction: column;
+    display: none;
+    top: 15px;
+    right: 15px;
+    display: none;
+    cursor: pointer;
+    position: absolute;
+    z-index: 1001;
+   
+}
+.hamburger-menu span{
+    width: 25px;
+    height: 3px;
+    margin: 3px 0;
+    background-color: white;
+    transition: all 0.3s ease;
 
+}
+.hamburger-menu.active span:nth-child(1) {
+  transform: rotate(45deg) translate(7px, 7px);
+}
+  .hamburger-menu.active span:nth-child(2) {
+    opacity: 0;
+  }
+  .hamburger-menu.active span:nth-child(3) {
+    transform: rotate(-45deg) translate(6px, -6px);
+  }
 
- @media screen and (min-width:280px) and (max-width: 319px){
+  @media screen and (min-width:280px) and (max-width: 319px){
     .show-nav{
         display: none;
         flex-direction: column;
@@ -49,7 +72,6 @@ span{
         align-items: center;
         width: 100%;
         background-color: rgb(41,67,76);
-        margin-left: -15px;
         top: 0;
         gap: 20px;
         z-index: 1000;
@@ -75,40 +97,13 @@ span{
         margin-bottom: 50px;
     
     }
-    span{
-        font-size: 65px;  
-    }
-    #landingPage{
-        background-image: url('/public/media/recycling.png');
-        background-color: rgb(41,67,76);
-        background-repeat: no-repeat;
-        background-position:  50% 50%;
-        background-size: 130%,130%;
-        height: 500px;
-        width: 100%;
-        margin: 0;
-        top: 0;
-        width: 100%;
-        
-    }
-  
     .Navbar-container{
       flex-direction: row;
       justify-content: space-between;
-      margin-left: 21px;
       margin-bottom: 40px;
      
     }
-    button{
-        border-radius: 30px;
-        border: none;
-        padding: 10px 32px;
-        text-align: center;
-        display: inline-block;
-        font-size: 20px;  
-        margin-top: 90px;   
-    }
- 
+  
 }
 
  
@@ -128,7 +123,7 @@ span{
         left: 0;
         height: 600px;
         position: fixed;
-        
+        margin-right: 20px;
       
     }
     .show-nav.active{
@@ -144,38 +139,15 @@ span{
         width: 150px;
         margin-bottom: 50px;
     }
-    span{
-        font-size: 70px;  
-    }
-    #landingPage{
-        background-image: url('/public/media/recycling.png');
-        background-color: rgb(41,67,76);
-        background-repeat: no-repeat;
-        background-position:  50% 50%;
-        background-size: 130%,130%;
-        height: 600px;
-        width: 100%;
-        margin: 0;
-        top: 0;
-        
-    }
-  
+   
     .Navbar-container{
       flex-direction: row;
       justify-content: space-between;
-      margin-left: 21px;
+      /* margin-left: 21px; */
       margin-bottom: 40px;
      
     }
-    button{
-        border-radius: 30px;
-        border: none;
-        padding: 15px 32px;
-        text-align: center;
-        display: inline-block;
-        font-size: 20px;  
-        margin-top: 130px;   
-    }
+
  
 }
 
@@ -187,7 +159,6 @@ span{
         align-items: center;
         width: 100%;
         background-color: rgb(41,67,76);
-        margin-left: -25px;
         top: 0;
         gap: 20px;
         z-index: 1000;
@@ -205,23 +176,7 @@ span{
         display: flex;
        
     }
-    span{
-        font-size: 120px;  
-        
-    }
-    #landingPage{
-        background-image: url('/public/media/recycling.png');
-        background-color: rgb(41,67,76);
-        background-repeat: no-repeat;
-        background-position:  50% 50%;
-        background-size: 130%,130%;
-        height: 700px;
-        width: 100%;
-        margin: 0;
-        top: 0;
-        
-      
-    }
+  
     .logo-image{
         height: 130px;
         width: 170px;
@@ -231,20 +186,10 @@ span{
     .Navbar-container{
         flex-direction: row;
         justify-content: space-between;
-        margin-left: 21px;
+     
 
     }
-    button{
-        border-radius: 30px;
-        border: none;
-        padding: 20px 32px;
-        text-align: center;
-        display: inline-block;
-        font-size: 20px; 
-        margin-top: 200px;
-        
-    }
-    
+ 
  
 }
    
@@ -257,7 +202,6 @@ span{
         align-items: center;
         width: 100%;
         background-color: rgb(41,67,76);
-        margin-left: -15px;
         top: 0;
         gap: 20px;
         z-index: 1000;
@@ -276,20 +220,7 @@ span{
         margin-top: 20px;
 
     }
-    span{
-        font-size: 150px;  
-    }
-    #landingPage{
-        background-image: url('/public/media/recycling.png');
-        background-color: rgb(41,67,76);
-        background-repeat: no-repeat;
-        background-position:  50% 50%;
-        background-size: 130%,130%;
-        height: 700px;
-        width: 100%;
-        margin: 0;
-        top: 0;
-    }
+   
     .logo-image{
         height: 150px;
         width: 200px;
@@ -299,23 +230,11 @@ span{
     .Navbar-container{
       flex-direction: row;
       justify-content: space-between;
-      margin: 12px;
+
       
 
     }
-    button{
-        border-radius: 30px;
-        border: none;
-        padding: 30px 42px;
-        text-align: center;
-        display: inline-block;
-        font-size: 20px; 
-
-        margin-top: 10px;
-        
     
-          
-    }
 }
 @media only screen and (min-width: 821px) and (max-width: 1020px){
     .show-nav{
@@ -325,7 +244,6 @@ span{
         align-items: center;
         width: 100%;
         background-color: rgb(41,67,76);
-        margin-left: -15px;
         top: 0;
         gap: 20px;
         z-index: 1000;
@@ -354,21 +272,7 @@ span{
     .hamburger-menu{
         display: flex;
     }
-    span{
-        font-size: 150px;  
-    
-    }
-    #landingPage{
-        background-image: url('/public/media/recycling.png');
-        background-color: rgb(41,67,76);
-        background-repeat: no-repeat;
-        background-position:  50% 50%;
-        background-size: 100%,100%;
-        height: 1300px;
-        width: 100%;
-      
-    }
-
+   
     .logo-image{
         margin-bottom: 40px;
 
@@ -381,18 +285,7 @@ span{
         height: 200px;
      
     }
-    button{
-        border-radius: 30px;
-        padding: 20px 35px;
-        text-align: center;
-        font-size: 35px;  
-        margin-bottom: 80px;
-
-     
-    }  
-    p{
-        font-size: 40px;
-    } 
+  
     .hamburger-menu.active span:nth-child(1) {
         transform: rotate(45deg) translate(22px, 22px);
       }
